Improve integration test workflow
This change updates the integration test setup script
to pull the integration test code directly from GitHub.

It also adds a "run" argument to the setup script;
by default the command will perform all setup steps,
but not run the test. This is useful in the development
case where you are modifying the integration test.