Add PubSubCleaner and DoPeriodically
This change adds the PubSubCleaner, which periodically
scans for unused projects and deletes the associated PubSub
and Spanner resources. It also adds unused project functionality
to the store.

It also adds the DoPeriodically function used for periodically
recurring tasks (such as the PubSubCleaner scan).

Finally, it fixes some test breakages in logentry and task
(not sure how these ever worked!)